Fix build + cleanup
